app/requirements.txt
python-telegram-bot
ujson
redis
